+++ docs/conf.py
@@ -56,7 +56,7 @@ author = "Scientific Software"
 # the built documents.
 #
 # The short X.Y version.
-version = "0.9.1"
+version = "0.10.0"
 # The full version, including alpha/beta/rc tags.
 release = version